Register IntentReceivers at the manifest level
This means that they will get triggered even when the app is dead. To do
this, we have to run a SongListener singleton that will  maintain the
current state of songs. This should be pretty golden.

This brings us down to only supporting google music right now, but we'll
soon abstract the rest.